added call_map_api
This function uses the GBIF map API to get a web map tile for a data set specified by the query arguments and converts the tile into a raster object. This can be used to quickly retrieve large amounts of aggregated data for analysis or for bias correction in species distribution mapping.