added call_map_api
This function calls the GBIF map API and uses the arguments passed on to the function to get a web map tile of the specified data. This map tile is converted into a raster file which can be used for diversity mapping, analyses, and bias correction in species distribution models.